requirements.txt
autopep8==1.5.5
click==7.1.2
cloudpickle==1.6.0
coloredlogs==15.0
dotmap==1.3.23
future==0.18.2
gym==0.18.0
humanfriendly==9.1
numpy==1.19.4
Pillow==7.2.0
pycodestyle==2.6.0
pyglet==1.5.0
PyYAML==5.4.1
scipy==1.5.2
toml==0.10.2
typer==0.3.2
tensorflow==2.4.1